# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lopes Signature real estate portal and administration API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: string;
  title: string;
  location: string;
  neighborhood: string;
  category: string;
  price: number;
  area: number;
  bedrooms: number;
  suites: number;
  parking: number;
  description: string;
  images: string[];
  badges: string[];
  featured: boolean;
  lpUrl: string;
  status: string;
}

export interface PropertyInput {
  title: string;
  location: string;
  neighborhood: string;
  category: string;
  price: number;
  area: number;
  bedrooms: number;
  suites: number;
  parking: number;
  description: string;
  images: string[];
  badges: string[];
  featured: boolean;
  lpUrl: string;
  status: string;
}

export interface AnalyticsSummary {
  totalVisits: number;
  whatsappClicks: number;
  lpViews: number;
  formConversions: number;
  conversionRate: number;
  visitsChange: number;
  leadsChange: number;
}

export interface AnalyticsPoint {
  label: string;
  visits: number;
  lpViews: number;
  leads: number;
}

export interface TrackingSettings {
  metaPixelId: string;
  metaEnabled: boolean;
  gtmContainerId: string;
  ga4MeasurementId: string;
  customHeadScript: string;
  customBodyScript: string;
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  new: 'new',
  contacted: 'contacted',
  scheduled: 'scheduled',
  negotiation: 'negotiation',
  closed: 'closed',
} as const;

export interface Lead {
  id: string;
  name: string;
  email: string;
  phone: string;
  propertyId: string;
  propertyTitle: string;
  status: LeadStatus;
  source: string;
  createdAt: string;
  note: string;
}

export type LeadInputStatus = typeof LeadInputStatus[keyof typeof LeadInputStatus];


export const LeadInputStatus = {
  new: 'new',
  contacted: 'contacted',
  scheduled: 'scheduled',
  negotiation: 'negotiation',
  closed: 'closed',
} as const;

export interface LeadInput {
  name: string;
  email: string;
  phone: string;
  propertyId: string;
  propertyTitle: string;
  status: LeadInputStatus;
  source: string;
  note: string;
}

export interface Specialist {
  id: string;
  name: string;
  role: string;
  credential: string;
  bio: string;
  image: string;
  whatsapp: string;
  listings: number;
}

export type NotFoundResponse = {
  message: string;
};

export type ListPropertiesParams = {
search?: string;
category?: string;
region?: string;
featured?: boolean;
};