# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lopes Signature real estate portal and administration API
 * OpenAPI spec version: 0.2.0
 */

export interface AnalyticsSummary {
  totalVisits: number;
  whatsappClicks: number;
  lpViews: number;
  formConversions: number;
  conversionRate: number;
  visitsChange: number;
  leadsChange: number;
}