# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lopes Signature real estate portal and administration API
 * OpenAPI spec version: 0.2.0
 */
import type { LeadInputStatus } from './leadInputStatus';

export interface LeadInput {
  name: string;
  email: string;
  phone: string;
  propertyId: string;
  propertyTitle: string;
  status: LeadInputStatus;
  source: string;
  note: string;
}