# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Lopes Signature real estate portal and administration API
 * OpenAPI spec version: 0.2.0
 */

export interface PropertyInput {
  title: string;
  location: string;
  neighborhood: string;
  category: string;
  price: number;
  area: number;
  bedrooms: number;
  suites: number;
  parking: number;
  description: string;
  images: string[];
  badges: string[];
  featured: boolean;
  lpUrl: string;
  status: string;
}